Add relaxed ordering settings and debug features for MLX5 driver
- Introduced `set_relaxed_ordering_write` and `set_relaxed_ordering_read` methods in `MkeyContextLayout` to configure relaxed ordering for memory keys.
- Enhanced `SendQueue` with a debug snapshot feature, allowing the recording of WQE states for better diagnostics.
- Implemented a mechanism to submit diagnostic TX frames during the initialization of the MLX5 bridge, ensuring proper logging and error handling.
- Added tests to validate the handling of large DHCP lease timers without overflow, ensuring robustness in lease time calculations.

+++ kernel/src/net/services/dhcp/client_impl/mod.rs
@@ -712,7 +712,10 @@ impl DhcpClient {
         current_tick: u64,
     ) -> DhcpLease {
         let t1 = opts.renewal_time.unwrap_or(opts.lease_time / 2);
-        let t2 = opts.rebinding_time.unwrap_or((opts.lease_time * 7) / 8);
+        // Promote to u64 before multiplying so large lease times do not panic in debug builds.
+        let t2 = opts
+            .rebinding_time
+            .unwrap_or(((opts.lease_time as u64 * 7) / 8) as u32);
 
         DhcpLease {
             ip_address: header.yiaddr(),
